refactor: search result feature - move directory & refactoring codes
