+++ src/components/SearchResult/SearchResultCard/index.ts
@@ -1,0 +1,3 @@
+import SearchResultCard from './SearchResultCard';
+
+export default SearchResultCard;
