+++ src/components/SearchResult/SearchResultCard/styles.ts
