+++ src/components/SearchResult/SearchResultCardList/index.ts
@@ -1,0 +1,3 @@
+import SearchResultCardList from './SearchResultCardList';
+
+export default SearchResultCardList;
