+++ src/components/SearchResult/SearchResultCardList/styles.ts
