+++ src/components/SearchResult/SearchResultHeader/index.ts
@@ -1,0 +1,3 @@
+import SearchResultHeader from './SearchResultHeader';
+
+export default SearchResultHeader;
