+++ src/components/SearchResult/SearchResultHeader/styles.ts
