+++ src/components/SearchResult/SearchResultInfinityScroll/index.ts
@@ -1,0 +1,3 @@
+import SearchResultInfinityScroll from './SearchResultInfinityScroll';
+
+export default SearchResultInfinityScroll;
