+++ src/components/SearchResult/SearchResultInfinityScroll/styles.ts
@@ -1,13 +1,5 @@
 import styled from 'styled-components';
 
-export const StyledSearchResultContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 20px;
-`;
-
 export const StyledTarget = styled.div`
   height: 1px;
 `;
