+++ src/components/SearchResult/SearchResultSkeleton/index.ts
@@ -1,0 +1,3 @@
+import SearchResultSkeleton from './SearchResultSkeleton';
+
+export default SearchResultSkeleton;
