+++ src/components/SearchResult/SearchResultSkeleton/styles.ts
