+++ src/pages/SearchPage/styles.ts
@@ -5,3 +5,11 @@ export const StyledSearch = styled.section`
   justify-content: center;
   padding-top: 16px;
 `;
+
+export const StyledSearchResultContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 20px;
+`;
