+++ src/utils/hooks/useSearchResult/index.ts
@@ -1,0 +1,3 @@
+import useSearchResult from './useSearchResult';
+
+export default useSearchResult;
